Rename some pages to avoid duplicating file names, and move some components into their own files
