+++ apps/frontend/components/Logo.js
@@ -1,0 +1,6 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faFire } from '@fortawesome/free-solid-svg-icons';
+
+export const Logo = () => {
+	return <FontAwesomeIcon icon={faFire} className={'text-orange-400 h-8'} />;
+};
